for #2428, SQLParserEngine => SQLParserEntry for parse SPI
