+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/parser/SQLParserFactory.java
@@ -21,7 +21,7 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.core.constant.DatabaseType;
 import org.apache.shardingsphere.core.parse.api.SQLParser;
-import org.apache.shardingsphere.core.parse.spi.SQLParserEngine;
+import org.apache.shardingsphere.core.parse.spi.SQLParserEntry;
 import org.apache.shardingsphere.core.spi.NewInstanceServiceLoader;
 
 /**
@@ -34,7 +34,7 @@ import org.apache.shardingsphere.core.spi.NewInstanceServiceLoader;
 public final class SQLParserFactory {
     
     static {
-        NewInstanceServiceLoader.register(SQLParserEngine.class);
+        NewInstanceServiceLoader.register(SQLParserEntry.class);
     }
     
     /** 
@@ -45,7 +45,7 @@ public final class SQLParserFactory {
      * @return SQL parser
      */
     public static SQLParser newInstance(final DatabaseType databaseType, final String sql) {
-        for (SQLParserEngine each : NewInstanceServiceLoader.newServiceInstances(SQLParserEngine.class)) {
+        for (SQLParserEntry each : NewInstanceServiceLoader.newServiceInstances(SQLParserEntry.class)) {
             if (DatabaseType.valueOf(each.getDatabaseType()) == databaseType) {
                 return each.createSQLParser(sql);
             }
