+++ sharding-core/sharding-core-parse/sharding-core-parse-oracle/src/main/resources/META-INF/services/org.apache.shardingsphere.core.parse.spi.SQLParserEntry
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.parse.SQLServerParserEngine
+org.apache.shardingsphere.core.parse.OracleParserEntry
