+++ sharding-core/sharding-core-parse/sharding-core-parse-postgresql/src/main/java/org/apache/shardingsphere/core/parse/PostgreSQLParserEntry.java
@@ -19,23 +19,23 @@ package org.apache.shardingsphere.core.parse;
 
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
-import org.apache.shardingsphere.core.parse.autogen.SQLServerStatementLexer;
-import org.apache.shardingsphere.core.parse.spi.SQLParserEngine;
+import org.apache.shardingsphere.core.parse.autogen.PostgreSQLStatementLexer;
+import org.apache.shardingsphere.core.parse.spi.SQLParserEntry;
 
 /**
- * SQL parser engine for SQLServer.
+ * SQL parser entry for PostgreSQL.
  *
  * @author zhangliang
  */
-public final class SQLServerParserEngine implements SQLParserEngine {
+public final class PostgreSQLParserEntry implements SQLParserEntry {
     
     @Override
     public String getDatabaseType() {
-        return "SQLServer";
+        return "PostgreSQL";
     }
     
     @Override
-    public SQLServerParser createSQLParser(final String sql) {
-        return new SQLServerParser(new CommonTokenStream(new SQLServerStatementLexer(CharStreams.fromString(sql))));
+    public PostgreSQLParser createSQLParser(final String sql) {
+        return new PostgreSQLParser(new CommonTokenStream(new PostgreSQLStatementLexer(CharStreams.fromString(sql))));
     }
 }
