+++ sharding-core/sharding-core-parse/sharding-core-parse-sqlserver/src/main/resources/META-INF/services/org.apache.shardingsphere.core.parse.spi.SQLParserEntry
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.parse.PostgreSQLParserEngine
+org.apache.shardingsphere.core.parse.SQLServerParserEntry
